refactor(maps): co-locate price filter helpers in map output
generatePriceNoteRegex, the input validators, and the small helpers used
only by the price filter move into OptimizedMapOutput.ts alongside
addQuantifier, qualityQualifier, and the existing optimizer; the dedicated
PriceNoteRegex module is no longer needed. optimize() is generalized to
collapse single-digit [X-Y] ranges (preserving the existing [8-9] and
[9-9] handling and replacing the price-side simplifySmallRanges). The
duplicate min/max validators are merged into one isValidPriceAmount.
